Tech debt audit: fix data races, harden driver, clean up dead code
Tier 1 (Safety):
- Make driver globals atomic (_Atomic for gRingBuffer, gVolume, gMute,
  gDevice_HostTicksPerFrame, gDevice_AnchorHostTime) with proper
  acquire/release ordering in IO paths
- Replace NSLock with os_unfair_lock in AudioMixer for RT-safe sync
- Add inDataSize validation in GetPropertyData/SetPropertyData to
  prevent buffer overflows
- Add inQualifierData NULL check before dereferencing

Tier 2 (Correctness):
- Add PropertiesChanged notifications for sample rate and stream
  format changes so HAL clients see updates
- Add @MainActor to AppState, remove listener in deinit
- Delete dead code: Views/, Permissions.swift, uninstallDriver()
- Call destroySharedMemory() on stop to clean up POSIX shm
- Replace all NSLog with os.Logger
- Persist user preferences via UserDefaults
- Fix MicCapture callback to pass channel count as UInt32
- Fix statusItem to NSStatusItem? with guard let
- Remove double dispatch in AppDelegate and AudioDeviceManager
- Add AudioConstants.defaultGain shared constant

Tier 3 (Build/CI):
- Add Debug/Release configurations to project.yml
- Enable hardened runtime on driver target
- Move kSupportedSampleRates to extern in header, define in .c
- Add _Static_assert on SharedRingBuffer struct layout
- Extract notarization to packaging/scripts/notarize.sh with API key auth
- Add generate.sh wrapper for xcodegen with post-gen recovery
- Add MIT LICENSE, .gitignore updates, test-pipeline.yml

Tier 4 (Polish):
- Reuse gLog instead of per-call os_log_create
- Extract recalcHostTicksPerFrame() helper
- Harden Installer scripts with set -euo pipefail, xattr, launchctl

Installer:
- Preinstall now stops running MacAudio app before replacing files

Tests:
- Add 189 tests (was 0): 20 C ring buffer, 81 pipeline, 88 source
  integrity covering all tiers

+++ MacAudio/AudioEngine/MicCapture.swift
@@ -10,7 +10,7 @@ final class MicCapture {
     private let logger = Logger(subsystem: "com.macaudio.app", category: "mic")
 
     /// Callback delivers interleaved Float32 frames at the mic's native sample rate
-    var onAudioReceived: ((UnsafePointer<Float>, UInt32, Float64) -> Void)?
+    var onAudioReceived: ((UnsafePointer<Float>, UInt32, UInt32) -> Void)?
 
     func setInputDevice(_ id: AudioDeviceID) {
         deviceID = id
@@ -72,7 +72,7 @@ final class MicCapture {
         guard frameCount > 0 else { return }
 
         let floatPtr = data.assumingMemoryBound(to: Float.self)
-        onAudioReceived?(floatPtr, frameCount, Float64(channelCount))
+        onAudioReceived?(floatPtr, frameCount, channelCount)
     }
 
     func stop() {
